refactor(lab/List): hide the focus owner channel behind a module-private symbol
