+++ src/components/lab/List/useListFocusOwner.ts
@@ -2,8 +2,12 @@
 
 import * as React from 'react';
 
+import {warnOnce} from '../../utils/warn';
+
 import {composeItemProps} from './composeItemProps';
-import type {ListFocusOwner, ListFocusOwnerConnection, ListFocusOwnerInputProps} from './types';
+import type {ListFocusOwnerConnection} from './focusOwnerChannel';
+import {LIST_FOCUS_OWNER_CHANNEL} from './focusOwnerChannel';
+import type {ListFocusOwner, ListFocusOwnerInputProps} from './types';
 
 interface OwnerState {
     listId?: string;
@@ -30,8 +34,16 @@ const DISCONNECTED: OwnerState = {};
  * ```
  *
  * `aria-expanded` считается по факту подключения списка: пока `<List>` с этим
- * владельцем не смонтирован, попап закрыт. Список, оставленный смонтированным
- * и спрятанный стилями, каналу неотличим от открытого.
+ * владельцем не смонтирован, попап закрыт. Канал рассчитан на mount/unmount-
+ * модель попапа — список, оставленный смонтированным и спрятанный стилями,
+ * каналу неотличим от открытого: стрелки продолжат двигать активность в
+ * невидимом попапе, а `aria-expanded` останется `true` (keepMounted не
+ * поддержан). Для не-попап паттернов (постоянно видимый фильтруемый список)
+ * `role`/`aria-expanded` переопределяются через overrides `getInputProps` —
+ * по контракту композиции последнее значение побеждает.
+ *
+ * Один владелец — один список: на два одновременно смонтированных списка
+ * нужно два объекта.
  */
 export function useListFocusOwner(): ListFocusOwner {
     const [state, setState] = React.useState<OwnerState>(DISCONNECTED);
@@ -43,22 +55,13 @@ export function useListFocusOwner(): ListFocusOwner {
     // Обработчик пересоздаётся ядром каждый рендер: если бы он лежал в
     // state, публикация зацикливалась бы на собственном ре-рендере
     const keyDownRef = React.useRef<ListFocusOwnerConnection['onKeyDown'] | null>(null);
+    // Синхронный след подключённого списка (state отстаёт на коммит) —
+    // только для dev-предупреждения о двух списках на одном владельце
+    const connectedListIdRef = React.useRef<string | undefined>(undefined);
 
     // Идентичность владельца стабильна: она входит в зависимости эффекта
     // публикации на стороне ядра
     const [owner] = React.useState<ListFocusOwner>(() => ({
-        connect({listId, activeDomId, onKeyDown}) {
-            keyDownRef.current = onKeyDown;
-            setState((prev) =>
-                prev.listId === listId && prev.activeDomId === activeDomId
-                    ? prev
-                    : {listId, activeDomId},
-            );
-        },
-        disconnect() {
-            keyDownRef.current = null;
-            setState((prev) => (prev === DISCONNECTED ? prev : DISCONNECTED));
-        },
         getInputProps(overrides) {
             const {listId, activeDomId} = stateRef.current;
             const baseProps = {
@@ -71,6 +74,31 @@ export function useListFocusOwner(): ListFocusOwner {
                 },
             };
             return composeItemProps(baseProps, overrides) as ListFocusOwnerInputProps;
+        },
+        [LIST_FOCUS_OWNER_CHANNEL]: {
+            connect({listId, activeDomId, onKeyDown}) {
+                if (
+                    process.env.NODE_ENV !== 'production' &&
+                    connectedListIdRef.current !== undefined &&
+                    connectedListIdRef.current !== listId
+                ) {
+                    warnOnce(
+                        '[List] Two mounted lists share one `useListFocusOwner()` object. The channel connects a single list at a time — the last connected list wins, and the other one silently loses the keyboard and aria wiring. Create a separate owner per list.',
+                    );
+                }
+                connectedListIdRef.current = listId;
+                keyDownRef.current = onKeyDown;
+                setState((prev) =>
+                    prev.listId === listId && prev.activeDomId === activeDomId
+                        ? prev
+                        : {listId, activeDomId},
+                );
+            },
+            disconnect() {
+                connectedListIdRef.current = undefined;
+                keyDownRef.current = null;
+                setState((prev) => (prev === DISCONNECTED ? prev : DISCONNECTED));
+            },
         },
     }));
